moving functions to contours module
